refactor(map): map features legacy 폴더로 이동
