+++ src/features/Map/legacy/constants.ts
@@ -1,4 +1,4 @@
-import type { MapTypeItem } from '@/features/map/model/types';
+import type { MapTypeItem } from '@/features/map/legacy/model/types';
 
 export const mapItemList: MapTypeItem[] = [
   { label: '일반', type: 'normal' },
